pyproject.toml
[project]
name = "cfd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celluloid>=0.2.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]
